+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -134,7 +134,7 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-02-21T15:10:20.378Z")
+    created: Optional[datetime] = parse_datetime("2026-02-23T02:47:38.871Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")