+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ0C15HPYY5QMK4MNYNNX1K7"
+        "01KJ46ARN1Q0G5XVDVJATZX99X"
     )
     r"""The unique identifier of the data source"""