+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -259,7 +259,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KJ0C15DEDC5CKHRDMR7E94VR"
+    id: Optional[str] = "01KJ46ARHV8DRDN12QW2ZSF32W"
 
     description: Optional[str] = None
 
@@ -1214,7 +1214,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KJ0C15D9CJV8SKTPQCR4XA6K"
+        "tool_01KJ46ARHCFHQGMJSNC7QYBQ16"
     )
 
     display_name: Optional[str] = None
@@ -1345,7 +1345,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KJ0C15D8B2P7DNE41DQAKHS8"
+    id: Optional[str] = "01KJ46ARHBDF3YRC3DJKB5HJT7"
 
     description: Optional[str] = None
 
@@ -1476,7 +1476,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KJ0C15D7E5SW557S8PFAG6ZC"
+        "tool_01KJ46ARHAADN3RZ8FDS3DAPES"
     )
 
     display_name: Optional[str] = None
@@ -1777,7 +1777,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KJ0C15D6H3RJR4HB7199ZXF7"
+        "tool_01KJ46ARH8YHY44Y83X4QVZ8BV"
     )
 
     display_name: Optional[str] = None
@@ -1972,7 +1972,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KJ0C15D4TVB9X2QBHGSP4RRG"
+        "tool_01KJ46ARH7PFD1DDWM7GWP0626"
     )
 
     display_name: Optional[str] = None
@@ -2171,7 +2171,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KJ0C15D22RX0FAGKES666B5Z"
+        "tool_01KJ46ARH63D965JV1SZEY62G4"
     )
 
     display_name: Optional[str] = None