+++ packages/orq-rc/RELEASES.md
@@ -5939,3 +5939,13 @@ Based on:
 - [python v4.4.0-rc.26] packages/orq-rc
 ### Releases
 - [PyPI v4.4.0-rc.26] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.26 - packages/orq-rc
+
+## 2026-02-23 02:50:21
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.723.0 (2.835.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.4.0-rc.27] packages/orq-rc
+### Releases
+- [PyPI v4.4.0-rc.27] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.27 - packages/orq-rc